scale funciona

+++ OS/Kernel/include/video.h
@@ -1,5 +1,5 @@
 #ifndef VIDEO_H
 #define VIDEO_H
 #include <stdint.h>
-void writeChar(uint8_t letter, unsigned int x, unsigned int y, unsigned int scale);
+void writeChar(uint8_t letter, unsigned int x, unsigned int y, double scale);
 #endif
